+++ src/main/java/com/api/dto/PayloadWeights.java
@@ -17,8 +17,8 @@ public class PayloadWeights {
     private String name;
 
     @JsonProperty("kg")
-    private long kg;
+    private Integer kg;
 
     @JsonProperty("lb")
-    private long lb;
+    private Integer lb;
 }
