+++ src/main/java/com/api/entity/SxThrustStage.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @Getter
 public class SxThrustStage {
     @JsonProperty("kn")
-    private int kn;
+    private Integer kn;
 
     @JsonProperty("lbf")
-    private long lbf;
+    private Integer lbf;
 }
